Changed all times to double

+++ engine/src/component/CameraComponent.hpp
@@ -33,7 +33,7 @@ public:
 
     void OnCreate();
     void OnDestroy();
-    void OnUpdate(float time_diff);
+    void OnUpdate(double time_diff);
 
     /**
       * Rotates the camera to look at a target point.
