+++ engine/src/event/BeginFrameEvent.cpp
@@ -10,7 +10,7 @@
 
 namespace dt {
 
-BeginFrameEvent::BeginFrameEvent(uint32_t frame_time) {
+BeginFrameEvent::BeginFrameEvent(double frame_time) {
     mFrameTime = frame_time;
 }
 
@@ -22,7 +22,7 @@ Event* BeginFrameEvent::Clone() const {
     return new BeginFrameEvent(mFrameTime);
 }
 
-uint32_t BeginFrameEvent::GetFrameTime() const {
+double BeginFrameEvent::GetFrameTime() const {
     return mFrameTime;
 }
